feat: Implementada funcionalidad de ver producto por ID

+++ src/controllers/product.controller.js
@@ -1,4 +1,7 @@
-import { ProductNotFoundError, productService } from "../services/product.service.js";
+import {
+  ProductNotFoundError,
+  productService,
+} from "../services/product.service.js";
 
 export class ProductController {
   /** @type {typeof productService} */
@@ -20,10 +23,25 @@ export class ProductController {
     const products = await this.#productService.findAllToSell();
 
     res.status(200).json({
-      products
+      products,
     });
   }
 
+  /**
+   * @param {import("express").Request}
+   * @param {import("express").Response}
+   */
+  async findById(req, res) {
+    const product_id = this.#parseProductId(req, res);
+    if (!product_id) return;
+    const product = await this.#productService.findById(product_id);
+    if (!product) {
+      return res.status(404).json({
+        message: "Producto no encontrado",
+      });
+    }
+    res.status(200).json({ product });
+  }
 
   /**
    * @param {import("express").Request}
@@ -35,12 +53,12 @@ export class ProductController {
       const products = await this.#productService.findByCategory(category);
 
       return res.status(200).json({
-        products
+        products,
       });
-    } catch(err) {
+    } catch (err) {
       if (err instanceof ProductNotFoundError) {
         return res.status(404).json({
-          message: err.message
+          message: err.message,
         });
       }
     }
@@ -55,7 +73,7 @@ export class ProductController {
     const products = await this.#productService.findBySellerId(user.user_id);
 
     res.status(200).json({
-      products
+      products,
     });
   }
 
@@ -71,8 +89,8 @@ export class ProductController {
     );
 
     return res.json({
-        product,
-        message: "Producto registrado exitosamente"
+      product,
+      message: "Producto registrado exitosamente",
     });
   }
 
@@ -83,24 +101,29 @@ export class ProductController {
   async updateProduct(req, res) {
     const user = req.user;
     const numberId = this.#parseProductId(req, res);
+    if (!numberId) return;
 
     try {
-      const updated = await this.#productService.update(numberId, req.body, user);
+      const updated = await this.#productService.update(
+        numberId,
+        req.body,
+        user
+      );
 
       return res.status(200).json({
         product: updated,
-        message: "Producto actualizado exitosamente"
+        message: "Producto actualizado exitosamente",
       });
-    } catch(err) {
+    } catch (err) {
       if (err instanceof ProductNotFoundError) {
         return res.status(404).json({
-          message: "Producto no encontrado"
+          message: "Producto no encontrado",
         });
       }
       console.error("Error al actualizar producto: ", err);
       return res.status(500).json({
-        message: "Error interno del servidor"
-      })
+        message: "Error interno del servidor",
+      });
     }
   }
 
@@ -108,9 +131,10 @@ export class ProductController {
     const product_id = req.params.product_id;
     const numberId = parseInt(product_id);
     if (!product_id || Number.isNaN(numberId)) {
-      return res.status(400).json({
-        message: "ID de producto no válido"
+      res.status(400).json({
+        message: "ID de producto no válido",
       });
+      return null;
     }
     return numberId;
   }
@@ -122,17 +146,18 @@ export class ProductController {
   async deleteProduct(req, res) {
     const user = req.user;
     const numberId = this.#parseProductId(req, res);
+    if (!numberId) return;
 
     const deleted = await this.#productService.delete(numberId, user);
 
     if (!deleted) {
       return res.status(404).json({
-        message: "Producto no encontrado"
-      })
+        message: "Producto no encontrado",
+      });
     }
 
-    return res.status(200).json({ 
-      message: "Producto eliminado exitosamente"
+    return res.status(200).json({
+      message: "Producto eliminado exitosamente",
     });
   }
 }
